+++ NEWS.md
@@ -1,5 +1,5 @@
 # gsignal 0.3.6
-- date: 20230830
+- date: 20240822
 - Fixed Github Issue #9: differences between freqz and signal.scipy.freqz
   * updated freqz() to match current Octave version
   * implemented proper freqz.Sos() instead of converting to Arma
@@ -21,6 +21,8 @@
 - Bugfix in fir1(): adapted calculation of w_o
 - Fixed Github Issue #19 (jefferis): changed control flow logic in findpeaks()
   (line 174)
+- Adapt freqs() to match freqz() - delete freqs_plot.R
+
 
 
 # gsignal 0.3-5